[WIP] Initial import of dtc repo for MultiROM.
This repository adds an Android-build system compatible
version of the dtc tool and the libfdt library. Source:
https://git.kernel.org/pub/scm/utils/dtc/dtc.git
This library enables the reading and manipulation of FDTs
appended to a kernel or boot image.

In the context of multirom, the libfdt library is used to
manipulate the FDT such that fstab entries can be removed
if needed. See the multirom repository for a more detailed
explanation.
TODO: link to mrom commit.

TO BE DISCUSSED:
* Is a shared lib variant needed? (e.g., for host binaries)

Signed-off-by: Alexander Diewald <Diewi@diewald-net.com>